requirements.txt
python-3.10
django-3.2.12